requirements.txt
fastapi
sqlmodel
uvicorn[standard]
pydantic[email]
pydantic-settings
alembic
python-dotenv
pytest
httpx
lxml
python-jose
argon2-cffi
python-multipart
click
aiohttp
sse-starlette
jinja2
deepseek
langchain
langchain-core
langchain-community
langchain-deepseek
langgraph
sqlite-utils
websockets